rename to datasets to include peoples art dataset
